Decouple the InputHandler class from EntryHandler class
