+++ src/main/java/seedu/wildwatch/operation/EntryHandler.java
@@ -5,11 +5,8 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import seedu.wildwatch.command.AddCommand;
-import seedu.wildwatch.command.ByeCommand;
 import seedu.wildwatch.command.DeleteCommand;
-import seedu.wildwatch.command.HelpCommand;
 import seedu.wildwatch.command.ListCommand;
-import seedu.wildwatch.entry.EntryList;
 import seedu.wildwatch.exception.IncorrectInputException;
 import seedu.wildwatch.exception.UnknownInputException;
 
@@ -17,38 +14,6 @@ import seedu.wildwatch.exception.UnknownInputException;
 public class EntryHandler {
     private static final int DEFAULT_NUMBER_INPUT = -3710; //Number that can never be input in normal use of WildWatch
     private static final Logger LOGGER = Logger.getLogger(EntryHandler.class.getName());
-
-    public static void handleManualInput() {
-        while (true) {
-            String inputBuffer = Ui.inputRetriever(); //Retrieves input of user
-            LOGGER.log(Level.INFO, "Input received: {0}", inputBuffer);
-
-            if (inputBuffer.equals("bye")) {        //Program exit
-                break;
-            } else if (inputBuffer.equals("help")) {  //User request "help"
-                Ui.printHorizontalLines();
-                Ui.helpRequestMessagePrinter();
-                Ui.printHorizontalLines();
-                HelpCommand.printHelpMessage();
-            } else {
-                Ui.printHorizontalLines();
-                ErrorHandler.handleError(inputBuffer);
-                Ui.printHorizontalLines();
-            }
-            EntryList.saveEntry();
-        }
-        ByeCommand.exitProgram();
-    }
-
-    public static void handleFileInput(String lineOfFile) {
-        try {
-            EntryHandler.handleEntry(lineOfFile, true);
-        } catch (UnknownInputException | IncorrectInputException exception) {
-            Ui.corruptFileMessagePrinter();
-            ShutDown.shutDown();
-            System.exit(0);
-        }
-    }
 
     public static void handleEntry(String inputBuffer, boolean isFromFile)
             throws UnknownInputException, IncorrectInputException {
